requirements.txt
mysql-connector-python==9.0.0
pytest-mock==3.14.0
pytest==8.3.3
